+++ server/core_account.go
@@ -21,6 +21,7 @@ import (
 	"errors"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/gofrs/uuid"
 	"github.com/heroiclabs/nakama-common/api"
@@ -469,7 +470,9 @@ func ExportAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID u
 	return export, nil
 }
 
-func DeleteAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID uuid.UUID, recorded bool) error {
+func DeleteAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, leaderboardRankCache LeaderboardRankCache, userID uuid.UUID, recorded bool) error {
+	ts := time.Now().UTC().Unix()
+
 	tx, err := db.BeginTx(ctx, nil)
 	if err != nil {
 		logger.Error("Could not begin database transaction.", zap.Error(err))
@@ -486,7 +489,7 @@ func DeleteAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID u
 			return nil
 		}
 
-		err = LeaderboardRecordsDeleteAll(ctx, logger, tx, userID)
+		err = LeaderboardRecordsDeleteAll(ctx, logger, leaderboardRankCache, tx, userID, ts)
 		if err != nil {
 			logger.Debug("Could not delete leaderboard records.", zap.Error(err), zap.String("user_id", userID.String()))
 			return err